ignite-3477-idx2 back to fixed size

+++ modules/indexing/src/main/java/org/apache/ignite/internal/processors/query/h2/database/FastIndexHelper.java
@@ -83,7 +83,7 @@ public class FastIndexHelper {
     /**
      * @return Value size.
      */
-    public int size() {
+    public short size() {
         switch (type) {
             case Value.BOOLEAN:
             case Value.BYTE:
